requirements.txt
mysqlclient
python-dotenv
djangorestframework
djangorestframework-simplejwt
stripe
Pillow
pandas
pytz

pytest-mock
